+++ internal/db/service/common_test.go
@@ -33,93 +33,93 @@ func setupTestDB(t *testing.T) (*gorm.DB, func()) {
 }
 
 // Helper functions to get type IDs from the database
-func getRegisteredModelTypeID(t *testing.T, db *gorm.DB) int64 {
+func getRegisteredModelTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.RegisteredModelTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find RegisteredModel type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getModelVersionTypeID(t *testing.T, db *gorm.DB) int64 {
+func getModelVersionTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.ModelVersionTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find ModelVersion type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getModelArtifactTypeID(t *testing.T, db *gorm.DB) int64 {
+func getModelArtifactTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.ModelArtifactTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find ModelArtifact type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getDocArtifactTypeID(t *testing.T, db *gorm.DB) int64 {
+func getDocArtifactTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.DocArtifactTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find DocArtifact type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getServingEnvironmentTypeID(t *testing.T, db *gorm.DB) int64 {
+func getServingEnvironmentTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.ServingEnvironmentTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find ServingEnvironment type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getInferenceServiceTypeID(t *testing.T, db *gorm.DB) int64 {
+func getInferenceServiceTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.InferenceServiceTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find InferenceService type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getServeModelTypeID(t *testing.T, db *gorm.DB) int64 {
+func getServeModelTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.ServeModelTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find ServeModel type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getExperimentTypeID(t *testing.T, db *gorm.DB) int64 {
+func getExperimentTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.ExperimentTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find Experiment type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getExperimentRunTypeID(t *testing.T, db *gorm.DB) int64 {
+func getExperimentRunTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.ExperimentRunTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find ExperimentRun type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getDataSetTypeID(t *testing.T, db *gorm.DB) int64 {
+func getDataSetTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.DataSetTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find DataSet type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getMetricTypeID(t *testing.T, db *gorm.DB) int64 {
+func getMetricTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.MetricTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find Metric type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getParameterTypeID(t *testing.T, db *gorm.DB) int64 {
+func getParameterTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.ParameterTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find Parameter type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
 
-func getMetricHistoryTypeID(t *testing.T, db *gorm.DB) int64 {
+func getMetricHistoryTypeID(t *testing.T, db *gorm.DB) int32 {
 	var typeRecord schema.Type
 	err := db.Where("name = ?", defaults.MetricHistoryTypeName).First(&typeRecord).Error
 	require.NoError(t, err, "Failed to find MetricHistory type")
-	return int64(typeRecord.ID)
+	return typeRecord.ID
 }
